server/requirements.txt
Django==1.8
django-post-office==2.0.7
jsonfield==1.0.3
